% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer_core.R
\name{as.character.entrainer_response}
\alias{as.character.entrainer_response}
\title{Convert an EntraineR response to character}
\usage{
\method{as.character}{entrainer_response}(x, ...)
}
\arguments{
\item{x}{Object returned by gemini_generate().}

\item{...}{Unused.}
}
\value{
Generated text.
}
\description{
Convert an EntraineR response to character
}
