% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer_core.R
\name{trainer_core_detect_main_factors}
\alias{trainer_core_detect_main_factors}
\title{Detect main-effect factor names present in T-test lines (ignore interactions)}
\usage{
trainer_core_detect_main_factors(tt_lines)
}
\arguments{
\item{tt_lines}{Character vector.}
}
\value{
Character vector of factor names.
}
\description{
Detect main-effect factor names present in T-test lines (ignore interactions)
}
\examples{
trainer_core_detect_main_factors(c("A - a", "B - b:C - c"))
}
