% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer_core.R
\name{trainer_core_filter_ttest_by_factors}
\alias{trainer_core_filter_ttest_by_factors}
\title{Filter T-test lines by requested factors (main and/or interactions)}
\usage{
trainer_core_filter_ttest_by_factors(
  tt_lines,
  keep_factors = NULL,
  keep_intercept = TRUE
)
}
\arguments{
\item{tt_lines}{Character vector of T-test lines.}

\item{keep_factors}{Character vector of factor names or "A:B".}

\item{keep_intercept}{Logical; keep (Intercept) line.}
}
\value{
Filtered character vector.
}
\description{
Filter T-test lines by requested factors (main and/or interactions)
}
\examples{
trainer_core_filter_ttest_by_factors(c("(Intercept)", "A - a", "A - b:B - c"), "A", TRUE)
}
