% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer_core.R
\name{trainer_core_ttest_scope_msg}
\alias{trainer_core_ttest_scope_msg}
\title{Scope message for T-test section based on requested & found factors}
\usage{
trainer_core_ttest_scope_msg(t_test, requested, actually_shown)
}
\arguments{
\item{t_test}{User request vector.}

\item{requested}{Vector of normalized requested items.}

\item{actually_shown}{Vector from \code{trainer_core_actually_shown()}.}
}
\value{
Single character scope message.
}
\description{
Scope message for T-test section based on requested & found factors
}
\examples{
trainer_core_ttest_scope_msg(c("A"), c("A"), c("A"))
}
